refactor how renderers are drawn

+++ source/CompiledShader.cs
@@ -6,12 +6,12 @@ namespace Rendering.Vulkan
     public readonly struct CompiledShader : IDisposable, IEquatable<CompiledShader>
     {
         public readonly bool isInstanced;
-        public readonly uint vertexVersion;
-        public readonly uint fragmentVersion;
+        public readonly ushort vertexVersion;
+        public readonly ushort fragmentVersion;
         public readonly ShaderModule vertexShader;
         public readonly ShaderModule fragmentShader;
 
-        public CompiledShader(uint vertexVersion, uint fragmentVersion, ShaderModule vertexShader, ShaderModule fragmentShader, bool isInstanced)
+        public CompiledShader(ushort vertexVersion, ushort fragmentVersion, ShaderModule vertexShader, ShaderModule fragmentShader, bool isInstanced)
         {
             this.isInstanced = isInstanced;
             this.vertexVersion = vertexVersion;
